Structuring of test in luna-studio backend
