+++ libs/luna-empire/test/SpecUtils/Graph.hs
@@ -1,0 +1,93 @@
+module SpecUtils.Graph where
+
+import           Control.Lens                    (uses)
+import           Data.Coerce                     (coerce)
+import qualified Data.Map                        as Map
+import           Data.UUID                       (UUID)
+import           Data.UUID.V4                    (nextRandom)
+import           Empire.ASTOp                    (runASTOp)
+import qualified Empire.Commands.Graph           as Graph
+import qualified Empire.Data.BreadcrumbHierarchy as BH
+import qualified Empire.Data.Graph               as Graph
+import           Empire.Empire                   (Empire)
+import           Empire.Prelude
+import           LunaStudio.Data.Breadcrumb      (Breadcrumb (Breadcrumb), BreadcrumbItem (Arg, Definition, Lambda))
+import           LunaStudio.Data.Connection      (Connection)
+import           LunaStudio.Data.GraphLocation   (GraphLocation (GraphLocation))
+import           LunaStudio.Data.Node            (ExpressionNode, NodeId)
+import qualified LunaStudio.Data.Node            as Node
+import           LunaStudio.Data.NodeLoc         (NodeLoc (NodeLoc))
+import           LunaStudio.Data.NodeMeta        (NodeMeta (NodeMeta))
+import           LunaStudio.Data.Port            (InPort, InPortId, InPortIndex (Self), OutPort, OutPortId, Port (Port), PortState)
+import           LunaStudio.Data.PortRef         (AnyPortRef (InPortRef'), InPortRef (InPortRef), OutPortRef (OutPortRef))
+import qualified LunaStudio.Data.Position        as Position
+import           LunaStudio.Data.TypeRep         (TypeRep (TStar))
+
+
+infixl 5 |>
+(|>) :: GraphLocation -> NodeId -> GraphLocation
+(|>) (GraphLocation file bc) nid = GraphLocation file 
+    . coerce . (<> [Lambda nid]) $ coerce bc
+
+infixl 5 |>-
+(|>-) :: GraphLocation -> (NodeId, Int) -> GraphLocation
+(|>-) (GraphLocation file bc) it = GraphLocation file 
+    . Breadcrumb . (<> [uncurry Arg it]) $ coerce bc
+
+infixl 5 |>=
+(|>=) :: GraphLocation -> NodeId -> GraphLocation
+(|>=) (GraphLocation file bc) it = GraphLocation file 
+    . Breadcrumb . (<> [Definition it]) $ coerce bc
+
+mkUUID :: MonadIO m => m UUID
+mkUUID = liftIO nextRandom
+        
+mkAllPort :: Text -> PortState -> OutPort
+mkAllPort name = Port mempty name TStar
+
+mkSelfPort :: PortState -> InPort
+mkSelfPort = Port [Self] "self" TStar
+
+mkAliasPort :: PortState -> InPort
+mkAliasPort = Port mempty "alias" TStar
+
+outPortRef :: NodeId -> OutPortId -> OutPortRef
+outPortRef = OutPortRef . NodeLoc def
+
+inPortRef :: NodeId -> InPortId -> InPortRef
+inPortRef = InPortRef . NodeLoc def
+
+
+addNode :: GraphLocation -> Text -> Empire ExpressionNode
+addNode gl code = mkUUID >>= \nid -> Graph.addNode gl nid code def
+
+connectToInput :: GraphLocation -> OutPortRef -> InPortRef -> Empire Connection
+connectToInput loc outPort inPort = Graph.connect loc outPort (InPortRef' inPort)
+
+findNodeIdByName :: GraphLocation -> Maybe Text -> Empire (Maybe NodeId)
+findNodeIdByName = fmap2 (view Node.nodeId) .: findNodeByName
+
+findNodeByName :: GraphLocation -> Maybe Text -> Empire (Maybe ExpressionNode)
+findNodeByName gl name = findNode <$> Graph.getNodes gl where
+    filterNodes = filter (\n -> n ^. Node.name == name)
+    findNode nodes = case filterNodes nodes of
+        [n] -> Just n
+        _   -> Nothing
+
+
+mockNodesLayout :: GraphLocation -> Empire ()
+mockNodesLayout gl = getMarkersWithNodeId >>= mapM_ (uncurry mockNodeMeta) where
+    getMarkersWithNodeId = Graph.withGraph gl $ do
+        markers <- fmap fromIntegral . Map.keys 
+            <$> use (Graph.userState . Graph.codeMarkers)
+        foundNodeIds <- runASTOp . fmap catMaybes . forM markers $ \m ->
+            (m,) `fmap2` Graph.getNodeIdForMarker m
+        topLevelNodeIds <- uses
+            (Graph.userState . Graph.breadcrumbHierarchy)
+            BH.topLevelIDs
+        pure $ filter (\(_, nid) -> elem nid topLevelNodeIds) foundNodeIds    
+    mockNodeMeta marker nid = Graph.setNodeMeta gl nid $ NodeMeta
+        (Position.fromTuple (0, fromIntegral marker * 10))
+        False
+        mempty
+
